geolocation-service/requirements.txt
spyne==2.13.16
lxml
python-dotenv
